Created a moveit package for abby.
Moveit is ROS's new planning package, which should combine 2d nav
and arm navigation. Unfortunately, it doesn't seem to be particularly
stable. I might switch over to this if I decide to go to groovy when
it releases next week, but the fuerte code (compiled from source)
is buggy and very poorly documented.