Creates a script to move the arm to stow position.
The abby_arm_actions package will eventually contain an action server
to take commands such as "stow arm," "store object," etc. For now,
it has a script to stow the arm. This script should work, but the
arm navigation stack currently thinks it is in collision even when it
is not and refuses to plan a trajectory.

Also adds the stl file for the bin to the repository.

Also adds some launch files to the arm navigation package.