OOP'd the arm actions.
Stow arm is now a class with a main method that runs the class's
method to stow the arm.

Store object is now a class with a main method that runs the class's
method to drop an object in the bin. Also added the gripper to the
store object method.

Neither of these are yet tested on the robot, but both are syntactically
good.