Added camera calibration files to repo for safe keeping.
The yaml files in abby_config/config/camera should be in ~/.ros/camera_info
for correct kinect calibration. They are copied here as a backup.

It is possible to explicitly specify camera calibration file location
in launch files, but that would give these preference over the default
calibration if the kinect hardware is ever replaced. As such, I'm using
the default location (~/.ros/camera_info) and just keeping these as a
backup copy.