Fixes arm location.
The arm was offset by several cm in the x direction and a few mm in the
y and z direction. Corrections were made by taking the x and y coordinates
from the kinect and lidar respectively and by lowering the arm as far as
possible and measuring the distance from gripper tip to floor to correct
the z.
(All coordinates above in /base_link frame.)