Fixed kinect and arm frames.
Kinect now publishes frames for rgb and depth in the correct
place. (curse you willowgarage).

IRB-120 has been moved and rotated to reflect the fact that the rail it is
mounted to is not actually vertical. A more complete fix would be to rotate
the whole frame in the model to be correct, but this will do for now.